+++ Assets/Scripts/World/Inventory/ItemTypesData/PotionsData/HealthPotionItemData.cs
@@ -5,6 +5,6 @@ namespace World.Inventory.ItemTypesData.PotionsData
     [CreateAssetMenu(fileName = "HealthPotionItemData", menuName = "Data/Inventory Data/Potion/Health")]
     public class HealthPotionItemData : PotionItemData
     {
-        public int healthPercent;
+        public float healthPercent;
     }
 }
